+++ src/js/hooks/inboundV2/addItems/useInboundAddItemsColumns.jsx
@@ -79,7 +79,7 @@ const useInboundAddItemsColumns = ({
   );
 
   const handleDelete = async (row) => {
-    if (getValues('currentLineItems').find((item) => item.id && item.id === row?.original?.itemId)) {
+    if (getValues('currentLineItems').find((item) => item.id === row?.original?.itemId)) {
       await removeSavedRow(row?.original?.itemId);
     }
     removeRow(row.index);
